BxTree Work-in
Deleting out BPlusTree stuff that doesn't work.  Move Tests to Test folder.  Made BxTree.find() return a RecordPointer.  Fixed DDLParser and DMLParser to work with the new BxTree to create, insert, find, and delete values from the tree.  In StorageManager we now delete the index file in the db/Indexes folder when we drop a table.  Need to do serialize and deserialize for the indexes.